feat: added support for seeding key generation
Also made the encapsulation API be able to take in an RNG. Shouldn't be
breaking, the old API just defaults to `OsRng`. Quite a bit more
complexity in key generation now because we can't rely on old versions
of `rand`, so we're doing deterministic key generation mostly internally
and generating from seeds.